Implement Clone Project and Experiment
This is achieved by giving CardStatus a temporary_split flag.
If this flag is set true, the unit will use Split on its turn.

On the appropriate turns (9 and 10 if experiment, every turn if
project), a skill called "temporary_split" is used by the commander to
award this status to one active unit.

To be determined: Whether the newly-split unit can activate Blitz.